backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.4",
    "pyright[nodejs]>=1.1.407",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.0.1",
    "ruff>=0.14.4",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]
